+++ CHANGELOG.md
@@ -10,3 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CSRF protection for main form.
 - Health check for Docker.
 - Heroku application manifest.
+
+### Changed
+- Server options moved to separate struct.